% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisOrganizeData.R
\name{analysisOrganizeData}
\alias{analysisOrganizeData}
\title{Analysis Organization & Data Preparation}
\usage{
analysisOrganizeData(df, analySpec = list(), reports = c(0, 1, 2, 3))
}
\arguments{
\item{df}{Data frame}

\item{analySpec}{Specifications for analysis}

\item{reports}{Optional reports about parameters, layers and stations
[default = c(0,1,2,3)]}
}
\value{
Returns a list. Use dfr[["df"]] and dfr[["analySpec"]] to extract
  updated data frame and updated (or created) analySpec. analySpec is a list
  that includes the following objects:

  analyTitle       - Analysis trend title

  parameterFilt    - Parameter filter used for down selection

  stationFilt      - Station filter used for down selection

  dateFilt         - Date filter for down selection

  setTZ            - time zone

  layerFilt        - Layer filter

  avgTechnique     - Averaging technique used to average layers (mean or median)

  layerAggOption   - Layer averaging option (see Details for more information)

  obsMin           - Minimum number of observations required to allow GAM analysis to proceed

  gamAlpha         - Alpha level used GAM analyses (where needed)

  censorTrim       - Values to apply for trimming data due to too much censoring

  gamModels        - model formulations

  gamDiffPeriods   - time periods (years) used for computing changes (differences)

  gamDiffSeasons   - seasons used for sub-annual analyses

  gamPenalty       - allow the user to set the mgcv::gam select argument to TRUE, FALSE, or baytrend algorithm default

  gamPenaltyCrit   - edf and F-stat values used to flag ANOVA table results

  gamCoeffDeltaMaxCrit - convergence criteria for expectation maximization

  gamFlw_Sal.Wgt.Perc - percentiles of flow [or salinity] to use for computing flow[salinity] averaged result

  gamLegend        - default settings for gam figure legend

  idVar            - primary key for data frame returned as df

  depVarList       - data frame of dependent variables (useful for setting up gam analyses in for loops)

  stationList      - data frame of stations (useful for setting up gam analyses in for loops)

  layerList        - data frame of layers (useful for setting up gam analyses in for loops)
}
\description{
This function assesses the user supplied specifications and prepares data for
analysis. In those cases where the user doesn't supply a needed
specification, a basic option is supplied by this function.
}
\details{
The following steps are performed:

 1) Review user supplied specifications. Fill in with default values. For
 example, if the user doesn't specify a list of stations, then all stations
 identified in the data set stationMasterList are used. Some other default values
 include the following: date range (1/1/1984-present), parameter list (all
 parameters in data set parameterList), layers (all layers in data set
 layerLukup), layer averaging technique ('mean'), layer aggregation option
 (0, no aggregation), minimum number of observations (60), GAM formulas
 (Linear Trend with Seasonality, Non-linear Trend with
 Seasonality, and Non-linear trend with Seasonality (plus Interactions)), GAM
 alpha level for plots and output (0.05), periods for multi-time period
 analyses using GAM (Full Record, 1999/00-Present, and 2005/06-Present), and
 seasons for sub-annual analyses using GAM (All, Spring1, Spring2, Summer1,
 Summer2, SAV1, SAV2, Winter, and Fall.)

 2) Down select primary data set based on parameters, stations, dates, and layers
 from updated specifications. A dependent variable list (depVarList) is
 created that includes variable descriptions, units, and log transformation
 selections. A station list (stationList) is created that includes the
 station ID, a selected USGS gage for correlating flow, and
 latitude/longitude.

 3) Aggregate data layers. analySpec$layerAggOption <- 0 results in no
 aggregation. The option analySpec$layerAggOption <- 1 would result in
 combining surface and above pycnocline data. In this example, records with
 layer = "S" and layer = "AP" are relabeled as "SAP". These records are then
 averaged to the date level. The averaging method (mean or median) is
 specified in by analySpec$avgTechnique. Other layerAggOption values are 2,
 3, 4, and 5 which average "B"&"BP"; "S"&"AP" and "B"&"BP"; all layers; and
 "S"&"B", respectively. A layer list (layerList) is created and returned.

 4) Date features are added. Columns for year, day of year (doy), decimal
 year (dyear), and month are added based on date.

 5) Reports on the number of records (0), parameters (1), layers (2) and
 stations (3) can be controlled with the reports option.
}
\examples{
# run analysis relying on default specifications, examine analySpec for
# default options
dfr <- analysisOrganizeData(dataCensored)
df        <- dfr[["df"]]
analySpec <- dfr[["analySpec"]]

# analyze bottom dissolved oxygen at 2 stations
analySpec <-list()
analySpec$parameterFilt <- c('do')
analySpec$layerFilt     <- c('B')
analySpec$stationFilt   <- c('CB3.3C', 'CB5.1')
dfr <- analysisOrganizeData(dataCensored, analySpec)
df        <- dfr[["df"]]
analySpec <- dfr[["analySpec"]]

}
