% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_as.character.R
\name{as.character.lcens}
\alias{as.character.lcens}
\alias{as.character.mcens}
\alias{as.character.qw}
\title{Character Representations}
\usage{
\method{as.character}{lcens}(x, ...)

\method{as.character}{mcens}(x, ...)

\method{as.character}{qw}(x, ...)
}
\arguments{
\item{x}{any R object to coerce to character}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
Creates a character representation of \code{x}.
}
\description{
Creates character representations of censored and water-quality data: 
methods for "lcens," "mcens," and "qw" data.  
Removed ATimport smwrBase (after keywords and before export)
}
\keyword{internal}
\keyword{manip}
