% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_as.data.frame.R
\name{as.data.frame.lcens}
\alias{as.data.frame.lcens}
\alias{as.data.frame.mcens}
\alias{as.data.frame.qw}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{lcens}(x, row.names = NULL, optional = FALSE,
  ..., nm = deparse(substitute(x)))

\method{as.data.frame}{mcens}(x, row.names = NULL, optional = FALSE,
  ..., nm = deparse(substitute(x)))

\method{as.data.frame}{qw}(x, row.names = NULL, optional = FALSE, ...,
  nm = deparse(substitute(x)), expand = FALSE)
}
\arguments{
\item{x}{any R object to put into a data frame.}

\item{row.names}{NULL or a character vector giving the row names for the data
frame. Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, set column names to \code{nm}.}

\item{\dots}{additional arguments to be passed to or from methods.}

\item{nm}{set as column name if \code{optional} is \code{TRUE}.}

\item{expand}{create a data frame from all of the slots from the object
instead of creating a data frame containting the object.}
}
\value{
A data frame constructred from \code{x}.
}
\description{
Creates data frames from censored and water-quality data: 
methods for "lcens," "mcens," and "qw" data.
}
\seealso{
\code{\link{data.frame}}
}
\keyword{internal}
\keyword{manip}
