% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baytrends.R
\docType{package}
\name{baytrends-package}
\alias{baytrends}
\alias{baytrends-package}
\title{baytrends: Long Term Water Quality Trend Analysis}
\description{
The baytrends package was developed to enable users to evaluate long-term
trends in the Chesapeake Bay using a Generalized Additive Modeling (GAM)
approach. The model development includes selecting a GAM structure to
describe nonlinear seasonally-varying changes over time, incorporation of
hydrologic variability via either a river flow or salinity, the use of an
intervention to deal with method or laboratory changes suspected to impact
data values, and representation of left- and interval-censored data. This
approach, which is fully transferable to other systems, allows for Chesapeake
Bay water quality data to be evaluated in a statistically rigorous, yet
flexible way to provide insights to a range of management- and
research-focused questions. Methodology described in Murphy, RR, E Perry, 
J Harcum, and J Keisman  2019  A Generalized Additive Model approach to 
evaluating water quality: Chesapeake Bay case study.  Environmental Modelling
& Software, 118 (2019) 1-13. 
<https://doi.org/10.1016/j.envsoft.2019.03.027>.
}
\details{
This software program is preliminary or provisional and is subject 
to revision. This software program is for testing only, no warranty, 
expressed or implied, is made as to the accuracy and functioning of the
program and related program material nor shall the fact of distribution 
constitute any such warranty, and no responsibility is assumed in connection 
therewith. This software is provided 'AS IS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tetratech/baytrends}
  \item Report bugs at \url{https://github.com/tetratech/baytrends/issues}
}

}
\author{
\strong{Maintainer}: Erik W Leppo \email{Erik.Leppo@tetratech.com} (\href{https://orcid.org/0000-0001-8026-7474}{ORCID})

Authors:
\itemize{
  \item Rebecca Murphy (\href{https://orcid.org/0000-0003-3391-1823}{ORCID})
  \item Elgin Perry
  \item Jennifer Keisman (\href{https://orcid.org/0000-0001-6808-9193}{ORCID})
  \item Jon Harcum (\href{https://orcid.org/0000-0001-5910-5243}{ORCID})
}

}
