% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baytrends.R
\docType{package}
\name{baytrends}
\alias{baytrends}
\alias{baytrends-package}
\title{baytrends: Long Term Water Quality Trend Analysis}
\description{
Enables users to develop and apply general additive model (GAM) 
to long-term water quality monitoring data sets.
}
\section{baytrends functions}{

analysisOrganizeData
as.qw
as.qw-method
...
}

