% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_censMulticomp.test.R
\name{censMulticomp.test}
\alias{censMulticomp.test}
\title{Multiple Comparisons}
\usage{
censMulticomp.test(x, groups, method = "holm", alpha = 0.05)
}
\arguments{
\item{x}{the observations the data are forced to class "lcens." Missing
values (NAs) are allowed and removed before the test is performed.}

\item{groups}{any group vector for the observations. Missing values (NAs) are
allowed and removed before the test is performed.}

\item{method}{the method for adjustment of p-value. See
\code{\link{p.adjust}} for options and details.}

\item{alpha}{the significance level of the test.}
}
\value{
An object of class "censMCT."
}
\description{
Performs a multiple nonparametric comparison test among groups of left-censored
data.
}
\note{
need some kind of general description.
}
\examples{

set.seed(69)
Xu <- rlnorm(22, 0, 1)
Yu <- rlnorm(22, .4, 1)
Zu <- rlnorm(22, 1, 1)
Gu <- factor(rep(c("X", "Y", "Z"), each=22))
# Censor the data at 1
censKSample.test(as.lcens(c(Xu, Yu, Zu), 1), Gu)
censMulticomp.test(as.lcens(c(Xu, Yu, Zu), 1), Gu)

}
\references{
Helsel, D.R. 2012, Statistics for censored environmental data 
using Minitab and R: New York, Wiley, 324 p.
}
\seealso{
\code{\link{censKSample.test}}
}
\keyword{censored}
\keyword{htest}
\keyword{internal}
\keyword{nonparametric}
