% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_censor.R
\name{censor}
\alias{censor}
\title{Multiply-censored Data Conversion}
\usage{
censor(x, left = -Inf, right = Inf)
}
\arguments{
\item{x}{a numeric vector. Missing values are permitted.}

\item{left}{a vector of left-censoring thresholds. If the length is greater
than 1, then they are replicated to the length of \code{x}.}

\item{right}{a vector of right-censoring thresholds. If the length is greater
than 1, then they are replicated to the length of \code{x}.}
}
\value{
An S4 object of class "lcens" if \code{right} is \code{Inf}, otherwise
an S4 object of class "mcens."
}
\description{
Censors a vector at user-set limits for left- or right-censoring.
}
\note{
Missing velues are generated wherever \code{right} is less than or equal
to \code{left}.\cr
The \code{censor} function is provided as a simple function to create either left-
or multiply-censored data from numeric values, primarily for simulation pruposes.
The functions \code{as.lcens} or \code{as.mcens} offer greater flexibility and 
should be used for real data.
}
\examples{
## Create a random vector
set.seed(651)
x.tmp <- rnorm(5)
censor(x.tmp, 0, .75)
}
\references{
Lorenz, D.L., in preparation
}
\seealso{
\code{\link{as.lcens}}, \code{\link{as.mcens}}
}
\keyword{manip}
\keyword{methods}
