% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_censQuantile.R
\name{censQuantile}
\alias{censQuantile}
\alias{censQuantile.default}
\alias{censQuantile.lcens}
\alias{censQuantile.mcens}
\alias{censQuantile.qw}
\alias{censQuantile.default}
\alias{censQuantile.lcens}
\alias{censQuantile.mcens}
\alias{censQuantile.qw}
\title{Sample Quantiles}
\usage{
censQuantile(x, probs = seq(0, 1, 0.25), na.rm = FALSE, method = "ROS",
  type = 2, alpha = 0.4)

\method{censQuantile}{default}(x, probs = seq(0, 1, 0.25), na.rm = FALSE,
  method = "", type = 2, alpha = 0.4)

\method{censQuantile}{lcens}(x, probs = seq(0, 1, 0.25), na.rm = FALSE,
  method = "ROS", type = 2, alpha = 0.4)

\method{censQuantile}{mcens}(x, probs = seq(0, 1, 0.25), na.rm = FALSE,
  method = "flipped K-M", type = 2, alpha = 0.4)

\method{censQuantile}{qw}(x, probs = seq(0, 1, 0.25), na.rm = FALSE,
  method = "log MLE", type = 2, alpha = 0.4)
}
\arguments{
\item{x}{an object of a censored-data class whose sample quantiles are
wanted.}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{na.rm}{logical; if \code{TRUE}, any \code{NA} and \code{NaN}s are
removed from \code{x} before the quantiles are computed.}

\item{method}{the method to use for computing quantiles. See \bold{Details}.}

\item{type}{an integer between 1 and 9 selecting one of the nine quantile
algorithms detailed below to be used. See \code{\link{quantile}} for a
description of the types.}

\item{alpha}{the offset fraction to be used, depending on \code{method};
typically in [0,1].}
}
\value{
An object of class "censQuantile" with one named component for each
level requested in \code{probs}. The values returned for 
\code{method}="flipped K-M" can be of class "lcens."
}
\description{
Computes sample or estimated quantiles corresponding to the given
probabilities.  The smallest observation (censored or uncensored) corresponds
to a probability of 0 and the largest to a probability of 1.
}
\details{
The methods available in the current version are "flipped K-M," "log ROS,"
"ROS," "log MLE," and "MLE." The method "flipped K-M" produces quantiles
using the Kaplan-Meier method on fliiped data described by Helsel (2012). The
methods "log ROS" and "log MLE" are described by Helsel, 2012 and Helsel and
Cohn (1988).  The methods "ROS" and "MLE" are similar to "log ROS" and "log
MLE" except that no log- and back-transforms are made on the data.
}
\examples{

set.seed(28)
Xu <- rnorm(23)
censQuantile(Xu) # Uncensored (true) quartiles
censQuantile(as.lcens(Xu, 0)) # How'd ROS do?
censQuantile(as.lcens(Xu, 0), method="MLE") # How'd ROS do?

}
\references{
Helsel, D.R. 2012, Statistics for censored environmental data 
using Minitab and R: New York, Wiley, 324 p.\cr
Helsel, D.R. and Cohn, T.A., 1988, Estimation of descriptive statistics for
multiply censored water quality data: Water Resources Research v. 24, n.
12, pp. 1997-2004
}
\seealso{
\code{\link{quantile}}, \code{\link{quantile.lcens}}
}
\keyword{univar}
