% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_censReg.R
\name{censReg}
\alias{censReg}
\title{Censored Data Regression}
\usage{
censReg(formula, data, subset, weights, na.action, dist = "normal")
}
\arguments{
\item{formula}{a formula describing the regression model. See \bold{Details}.}

\item{data}{the data to search for the variables in \code{formula}.}

\item{subset}{an expression to select a subset of the data.}

\item{weights}{a variable to use for weights.}

\item{na.action}{what to do with missing values.}

\item{dist}{the distribution of the data, either "normal," "lognormal,"
or "commonlog." See \bold{Details}.}
}
\value{
An object of class "censReg."
}
\description{
Builds a linear regression model for censored response data.
}
\details{
The left-hand side of the formula may be any numeric variable, just as with
\code{lm} or a variable of class "lcens," "mcens," or "qw."\cr
For un- or left-censored data, AMLE is used unless weights are specified in
the model, then MLE is used, through a call to \code{survreg}. For any other 
censored data, MLE is used.\cr
If \code{dist} is "normal," then the regression analysis assumes that the
residuals are normally distributed. If \code{dist} is "lognormal," then
the regression analysis assumes that the residuals are lognormally
distributed. In this case, predicted values are back transformed and
optionally bias corrected to represent the expected mean. If \code{dist}
is "commonlog," then the response data are transformed using \code{log10}
and those residuals are assumed to be normally distributed. No back 
transformation is made for predicted values.
}
\note{
A special feature of \code{censReg} is the ability to expand
the formula passed as a variable rather than as an expressly entered
formula. This feature is intended to facilitate certain kinds of scripts that 
construct a formula from combinations of variables in a dataset.

Objects of class "censReg" have these methods:
\describe{
\item{coef}{extract the regression coefficents}
\item{extractAIC}{compute another information criterion (AIC) value}
\item{fitted}{extract the fitted values}
\item{logLik}{compute the log-likelihood of the fitted model}
\item{nobs}{extract the number of observations used to fit the model}
\item{plot}{create some basic diagnostic plots}
\item{predict}{predict new values from the model}
\item{print}{print the model}
\item{residuals}{extract the residuals from the model}
\item{rmse}{compute the root mean squared error (residual standard error)}
\item{summary}{compute summary information about the model}
\item{vcov}{extract the variance-covariance matrix from the model}
\item{vif}{compute the variance inflation factors for the explanatory variables in the model}
}

The AMLE method has arbitrary limitations on the size of the data: 24 explanatory variables
and 5000 observations. Exceeding these will generate a warning message with the error code in
the output object.
}
\examples{

set.seed(345)
X <- runif(24, 1, 5)
Y <- X/2 + rnorm(24)
lm(Y ~ X) # the uncensored regression
censReg(as.lcens(Y, 1) ~ X) # censored at 1

}
\references{
Lorenz, 2015, smwrQW.\cr
Breen, R., 1996, Regression models: censored, sample selected, or truncated data:
Sage University Paper series on Quantitative Applications in the Social Sciences, 
07-111, Thousand Oaks, CA, \cr
Cohn, T.A., 1988, Adjusted maximum likelihood estimation of moments
of lognormal populations from type I censored samples: 
U.S. Geological Survey Open-File Report 88-350, 34 p.\cr
}
\seealso{
\code{\link{lm}}, \code{\link{survreg}}
}
\keyword{censored}
\keyword{regression}
