% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_censStats.R
\name{censStats}
\alias{censStats}
\alias{censStats.default}
\alias{censStats.lcens}
\alias{censStats.qw}
\alias{censStats.mcens}
\alias{censStats.default}
\alias{censStats.lcens}
\alias{censStats.mcens}
\alias{censStats.qw}
\title{Mean and Standard Deviation}
\usage{
censStats(x, method = "MLE", na.rm = FALSE, alpha = 0.4)

\method{censStats}{default}(x, method = "MLE", na.rm = FALSE, alpha = 0.4)

\method{censStats}{lcens}(x, method = "MLE", na.rm = FALSE, alpha = 0.4)

\method{censStats}{mcens}(x, method = "MLE", na.rm = FALSE, alpha = 0.4)

\method{censStats}{qw}(x, method = "log MLE", na.rm = FALSE, alpha = 0.4)
}
\arguments{
\item{x}{an object of a censored-data class whose sample mean and standard
deviation are needed.}

\item{method}{the method to use for computing the statistics. See
\bold{Details}.}

\item{na.rm}{logical; if \code{TRUE}, any NA and NaN's are removed from
\code{x} before the quantiles are computed.}

\item{alpha}{the offset fraction to be used, depending on \code{method};
typically in [0,1].}
}
\value{
A list with one component for mean and sd. If "log" methods are
specified, then the meanlog amd meansd components are also included. The
values returned for \code{method}="flipped K-M" can be of class "mcens."
}
\description{
Computes the mean and standard deviation of censored data.
}
\details{
The methods available in the current version are "log MLE," "MLE," "log ROS,"
"ROS," "log AMLE," "AMLE," and "flipped K-M." The method "flipped K-M"
produces statistics using the Kaplan-Meier method on flipped data described
by Helsel (2012). The methods "log ROS," "log MLE," and "log AMLE" are
described by Helsel (2012) and Helsel and Cohn (1988).  The methods "ROS,"
"MLE," and "AMLE" are similar to the previous except that no log- and
back-transforms are made on the data. For "log" methods, if any value in
\code{x} is negative, then a warning is generated and the returned list 
contains missing values.
}
\examples{

set.seed(936)
Y <- rlnorm(32)
# Uncensored statistics
censStats(Y, method="log MLE")
# Censored statistics, at 1 (37.5 percent)
censStats(as.lcens(Y, 1), method="log MLE")

}
\references{
Helsel, D.R. 2012, Statistics for censored environmental data 
using Minitab and R: New York, Wiley, 324 p.

Helsel, D.R. and Cohn, T.A., 1988, Estimation of descriptive statistics for
multiply censored water quality data: Water Resources Research v. 24, n.
12, pp.1997--2004
}
\seealso{
\code{\link{mdlAMLE}}, \code{\link{mdlKM}}, \code{\link{mdlKMstats}}, 
\code{\link{mdlMLE}}, \code{\link{mdlROS}}
}
\keyword{univariate}
