% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_censSumStats.R
\name{censSumStats}
\alias{censSumStats}
\title{Summary Statistics}
\usage{
censSumStats(..., group = NULL, Nums = TRUE, Stats = TRUE,
  Probs = (0:4)/4, na.rm = TRUE, method = "log MLE")
}
\arguments{
\item{\dots}{any number of vectors or a data frame of the data to
be summarized. If \code{\dots} is a data frame, then only numeric data,
including "lcens," "mcens," and "qw" data types, will be summarized.}

\item{group}{the data to group the data in \code{\dots}.}

\item{Nums}{logical, include the number of observations and numbers of 
censored values in the output data.frame.}

\item{Stats}{logical, include the mean and standard deviation
in the output data.frame.}

\item{Probs}{the argument to quantile specifying the desired probabilities
in the output data.frame}

\item{na.rm}{logical, remove missing values before processing?}

\item{method}{the method to use for computing the statistics. Must be one of
"log MLE," "MLE," "log ROS," "ROS," or "flipped K-M." The method "flipped K-M"
is only applicaple to un- or left-censored data.}
}
\value{
A data.frame with the selected statistcs for each variable
in \code{\dots} grouped by each unique value in \code{group}.
}
\description{
Computes selected summary statistics for censored data.
}
\keyword{univar}
