% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeOut.R
\name{closeOut}
\alias{closeOut}
\title{Document Processing Time and Other Session Time}
\usage{
closeOut(timeProcess = TRUE, contInfo = TRUE, sessInfo = TRUE)
}
\arguments{
\item{timeProcess}{Processing time}

\item{contInfo}{Contact information}

\item{sessInfo}{Session information}
}
\value{
Reports out processing time, contact information and session information
}
\description{
Document Processing Time and Other Session Time
}
\examples{
closeOut()
}
