% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_code01.R
\name{code01}
\alias{code01}
\alias{code01.default}
\alias{code01.data.frame}
\title{Code Censored Values}
\usage{
code01(..., criteria)

\method{code01}{default}(..., criteria)

\method{code01}{data.frame}(..., criteria)
}
\arguments{
\item{\dots}{either a data frame that contains columns of class "qw" or any combination of
individual vectors of class "integer," "numeric," "lcens," or "qw." Missing values are
removed before processing.}

\item{criteria}{a vector indicating the critical values for each repsective value in
\code{\dots}. May be a named vector containing critical values for the named argument
in \code{\dots} or not supplied. For any argument in \code{\dots} that does not have
a critical value, the critical value is the largest censoring level or if the data have
no censored values, the median. The critical value can never be smaller than
the largest censoring level and will be changed with a warning message. A special case
is for integer values that have a minimum of 0---for those data the default criterion is to
return 0 if the value is 0 and return 1 for any other value.}
}
\value{
A data frame containing the binary coded values. Each column has an attribute
called "critical.value" that reports the critical value used to recode the values in 
that column. The column names are taken from \code{\dots}; the rownames are derived 
from the sequential numbers of the original data.
}
\description{
Recodes censored data to a binary coding system with 0 indicating
less than some critical value and 1 indicating greater than or equal to that critical 
value.
}
