% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_codeU.R
\name{codeU}
\alias{codeU}
\alias{codeU.default}
\alias{codeU.data.frame}
\title{Code Censored Values}
\usage{
codeU(...)

\method{codeU}{default}(...)

\method{codeU}{data.frame}(...)
}
\arguments{
\item{\dots}{either a data frame that contains columns of class "qw" or any combination of
individual vectors of class "numeric," "lcens," or "qw." Missing values are
removed before processing.}
}
\value{
A data frame containing the u-coded values. The column names are taken 
from \code{\dots}; the rownames are derived from the sequential numbers of the 
original data.
}
\description{
Recodes censored data to u-scores, which is the sum of the sign of the 
differences beween each value and all other values. It is equivalent to the rank but 
scaled so the the mean is 0.
}
