% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_Compare-censored.R
\docType{methods}
\name{Compare-censored}
\alias{Compare-censored}
\alias{Compare,lcens,numeric-method}
\alias{Compare,mcens,numeric-method}
\alias{Compare,qw,numeric-method}
\alias{Compare,numeric,lcens-method}
\alias{Compare,numeric,mcens-method}
\alias{Compare,numeric,qw-method}
\title{Comparison Methods for \code{lcens}, \code{mcens}, and \code{qw} objects}
\usage{
\S4method{Compare}{lcens,numeric}(e1, e2)

\S4method{Compare}{mcens,numeric}(e1, e2)

\S4method{Compare}{qw,numeric}(e1, e2)

\S4method{Compare}{numeric,lcens}(e1, e2)

\S4method{Compare}{numeric,mcens}(e1, e2)

\S4method{Compare}{numeric,qw}(e1, e2)
}
\arguments{
\item{e1, e2}{numeric, censored, or water-quality data. Missing values are permitted in 
either argument and result in a missing value in the output.}
}
\value{
A logical vector the represents the result of the comparison between each
element in \code{e1} and \code{e2}.
}
\description{
Comprisons are possible and well-defined between numeric and
censored or water-quality data. Comparisons between censored data types are
not supported in this version.
}
\details{
For all comparisons, the censored data are converted to multiply censored data,
class "mcens." The conversion from class "qw" preserves the strict interpretation
of less-than values as greater than or equal to 0 and less than the reporting level.
The distinction of being strictly less than is lost for the conversion of left-censored
data, class "lcens."
}
\examples{
as.lcens(c(1, 3), 2) > 1
as.lcens(c(1, 3), 2) < 4

}
\keyword{internal}
\keyword{manip}
\keyword{methods}
