% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deSeasonalize.R
\name{deSeasonalize}
\alias{deSeasonalize}
\title{De-seasonalize variable}
\usage{
deSeasonalize(df, dep, deSeasonalizeModel = "doy", analySpec = analySpec,
  gamTable = FALSE, gamPlot = FALSE, flow.detrended = NA,
  salinity.detrended = NA)
}
\arguments{
\item{df}{data frame}

\item{dep}{variable}

\item{deSeasonalizeModel}{which gam formula is used. Default: 'doy'.}

\item{analySpec}{analytical specifications}

\item{gamTable}{gam table setting (set to FALSE to turn off table output)}

\item{gamPlot}{gam plot setting (set to FALSE to turn off plotting)}

\item{flow.detrended}{data generated by detrended.flow.  Default = NA.}

\item{salinity.detrended}{data generated by detrended.flow.  Default = NA.}
}
\value{
Returns a vector with results
}
\description{
Use GAM analysis to de-seasonalize variable. Relies on mgcv::gam to perform general additive model.
}
