% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrended.salinity.R
\name{detrended.salinity}
\alias{detrended.salinity}
\title{Detrended Salinty Data}
\usage{
detrended.salinity(df.sal, stations, yearStart, yearEnd, dvAvgWinSel = 30,
  lowess.f = 0.2, minObs = 40, sdOverRide = 10, dir.out = getwd())
}
\arguments{
\item{df.sal}{data frame with salinty data.}

\item{stations}{Stations}

\item{yearStart}{start year}

\item{yearEnd}{end year}

\item{dvAvgWinSel}{average window selection}

\item{lowess.f}{lowess}

\item{minObs}{Minimum number of observations (default is 40).}

\item{sdOverRide}{standard deviation override (default is 10)}

\item{dir.out}{Output directory for data file and plots.  Defaults to working directory.}
}
\value{
Creates a data file (salinity.RDA) and plots in the user specified directory.
Plots are also output to the screen.
}
\description{
Create salintiy data set from user data, detrend it, and plot it.
}
\details{
Input is a a user file with 4 columns (Station, Sample_Date, Layer, and salinity).
See structure of sal data in example below.
}
\examples{
# Show Example Dataset (sal)
str(sal)

# Define Function Inputs
df.sal        <- sal
stations      <- c("CB3.3C", "CB5.1", "CB5.4", "LE1.2", "LE3.1"
                         , "LE3.6", "TF2.2", "WT2.1") 
yearStart     <- 1984
yearEnd       <- 2016
dvAvgWinSel   <- c(30)
lowess.f      <- 0.2
minObs        <- 40
sdOverRide    <- 10
dir.out       <- file.path(getwd(), "data2")
                 
# Run Function
detrended.salinity(df.sal, stations, yearStart, yearEnd
                  , dvAvgWinSel, lowess.f, minObs, sdOverRide
                  , dir.out
                  )               
}
