% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendDateFeatures.R
\name{.appendDateFeatures}
\alias{.appendDateFeatures}
\title{Appends date features to data frame}
\usage{
.appendDateFeatures(df, var = "date")
}
\arguments{
\item{df}{data frame}

\item{var}{variable with stored date}
}
\value{
Returns data frame with appended date features including year, doy,
  decimal year, and month
}
\description{
Appends date features to data frame. Creates new column 'date' based on var
if date is not already in the data frame. The newly created (or existing)
date column is truncated to day. Columns for year, day of year (doy), decimal
year (dyear), and month are added based on date. This function relies on
smwrBase::baseDay and smwrBase::baseDay2decimal for doy and decimal year.
  
The baseDay and baseDay2decimal functions have been added to this package 
from smwrBase package.
}
\examples{

# retrieve Secchi depth for Station LE3.1, missing Secchi depth values are
# maintained (see value for secchi on 2/12/2008), no transformations are applied
dfr   <- selectData(dataCensored, 'secchi', 'LE3.1', 'S', transform=FALSE, missing=FALSE)
dfr1  <- dfr[[1]]
iSpec <- dfr[[2]]
# retrieve surface corrected chlorophyll-a concentrations for Station LE3.1,
# missing values are removed and transformation applied
dfr   <- selectData(dataCensored, 'chla', 'LE3.1', 'S')
dfr2  <- dfr[[1]]
iSpec <- dfr[[2]]

}
