% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkParameter.R
\name{.chkParameter}
\alias{.chkParameter}
\title{Reduce dataframe and parameter list based on user selected parameterFilt}
\usage{
.chkParameter(df, parameterFilt = parameterFilt, parameterList)
}
\arguments{
\item{df}{data frame}

\item{parameterFilt}{parameter filter}

\item{parameterList}{parameter list}
}
\value{
n/a
}
\description{
Reduce dataframe and parameter list based on user selected parameterFilt
}
\examples{
#df <- chkParameter(df,parameterFilt=c("tn", "tp"))
}
