% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectMaxFunctions.r
\name{.ExpLNmCens}
\alias{.ExpLNmCens}
\title{Expectation maximization function: Log-normal case, Cens}
\usage{
.ExpLNmCens(df, dep, mu, sigma)
}
\arguments{
\item{df}{data frame}

\item{dep}{dependent variable}

\item{mu}{predicted values from mgcv::gam}

\item{sigma}{model standard deviation}
}
\description{
Expectation maximization function: Log-normal case, Cens
}
\keyword{internal}
