% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectMaxFunctions.r
\name{.ExpNiCens}
\alias{.ExpNiCens}
\title{Expectation maximization function: Normal case, i censured}
\usage{
.ExpNiCens(l, u, mu, sigma, iCens = NA)
}
\arguments{
\item{l}{l}

\item{u}{u}

\item{mu}{predicted values from mgcv::gam}

\item{sigma}{model standard deviation}

\item{iCens}{default=NA}
}
\description{
Expectation maximization function: Normal case, i censured
}
\keyword{internal}
