% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectMaxFunctions.r
\name{.ExpNlCens}
\alias{.ExpNlCens}
\title{Expectation maximization function: Normal case, left censured}
\usage{
.ExpNlCens(l, u, mu, sigma, lCens = NA)
}
\arguments{
\item{l}{l}

\item{u}{u}

\item{mu}{predicted values from mgcv::gam}

\item{sigma}{model standard deviation}

\item{lCens}{default=NA}
}
\description{
Expectation maximization function: Normal case, left censured
}
\keyword{internal}
