% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectMaxFunctions.r
\name{.ExpNmCens}
\alias{.ExpNmCens}
\title{Expectation maximization function: Normal case}
\usage{
.ExpNmCens(df, dep, mu, sigma)
}
\arguments{
\item{df}{data frame}

\item{dep}{dependent variable}

\item{mu}{predicted values from mgcv::gam}

\item{sigma}{model standard deviation}
}
\description{
Expectation maximization function: Normal case
}
\keyword{internal}
