% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers2.R
\name{.F}
\alias{.F}
\title{Print out figure title  (customization of pandoc.emphasis and pandoc.strong )}
\usage{
.F(text, n = NULL, t = "e")
}
\arguments{
\item{text}{text of figure title}

\item{n}{figure number}

\item{t}{emphasis or stong}
}
\value{
n/a
}
\description{
Print out figure title  (customization of pandoc.emphasis and pandoc.strong )
}
\examples{
text<-"Hello World!"
.F(text)
.F(text, 4)
.F(text, 4,'e')
.F(text, 4,'s')
}
\seealso{
.F .H .H2 .H3 .H4 .P .T .V
}
\keyword{internal}
