% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamTables.R
\name{.gamANOVA}
\alias{.gamANOVA}
\title{Prepare ANOVA table for GAM analysis}
\usage{
.gamANOVA(gamo)
}
\arguments{
\item{gamo}{output from gam model}
}
\description{
Prepare ANOVA table for GAM analysis
}
\keyword{internal}
