% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamTables.R
\name{.gamDiffPORtbl}
\alias{.gamDiffPORtbl}
\title{Compute and present report on percent different for log-transformed data}
\usage{
.gamDiffPORtbl(por.diff, iSpec)
}
\arguments{
\item{por.diff}{Output from gam.por.diff}

\item{iSpec}{data set specifications}
}
\description{
Compute and present report on percent different for log-transformed data
}
