% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamPlotCalc.r
\name{.gamPlotCalc}
\alias{.gamPlotCalc}
\title{plots data and gam fit vs. time}
\usage{
.gamPlotCalc(
  dep,
  tsdat,
  pgam,
  iSpec,
  analySpec,
  t.deriv = FALSE,
  alpha = 0.05,
  dayStep = 10,
  step.pt = "none",
  q.doy,
  flow.detrended = flow.detrended,
  salinity.detrended = salinity.detrended
)
}
\arguments{
\item{dep}{variable dep}

\item{tsdat}{variable tsdat}

\item{pgam}{variable pgam}

\item{iSpec}{variable iSpec}

\item{analySpec}{variable analySpec}

\item{t.deriv}{variable t.deriv}

\item{alpha}{variable alpha}

\item{dayStep}{variable dayStep}

\item{step.pt}{variable step.pt}

\item{q.doy}{vector of days of year}

\item{flow.detrended}{data generated by detrended.flow.  Default = flow.detrended.}

\item{salinity.detrended}{data generated by detrended.salinity.  Default = salinity.detrended.}
}
\description{
plots data and gam fit vs. time
}
\keyword{internal}
