% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers2.R
\name{.H}
\alias{.H}
\title{Print out header (shortened pandoc.header)}
\usage{
.H(text, n = 1)
}
\arguments{
\item{text}{text of header}

\item{n}{header level number}
}
\value{
n/a
}
\description{
Print out header (shortened pandoc.header)
}
\examples{
.H("1st level header",1)
}
\seealso{
.F .H .H2 .H3 .H4 .P .T .V
}
\keyword{internal}
