% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers2.R
\name{.H1}
\alias{.H1}
\title{Print out 1st level header (shortened pandoc.header)}
\usage{
.H1(text)
}
\arguments{
\item{text}{text of header}
}
\value{
n/a
}
\description{
Print out 1st level header (shortened pandoc.header)
}
\examples{
.H1("1st level header")
.H3("3rd level header")
}
\seealso{
.F .H .H2 .H3 .H4 .P .T .V
}
