% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers2.R
\name{.H2}
\alias{.H2}
\title{Print out 2nd level header (shortened pandoc.header)}
\usage{
.H2(text)
}
\arguments{
\item{text}{text of header}
}
\value{
n/a
}
\description{
Print out 2nd level header (shortened pandoc.header)
}
\examples{
.H2("2nd level header")
.H3("3rd level header")
}
\seealso{
.F .H .H2 .H3 .H4 .P .T .V
}
\keyword{internal}
