% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeResults.r
\name{.initializeResults}
\alias{.initializeResults}
\title{####
Initialize stat.gam.result and chng.gam.result}
\usage{
.initializeResults()
}
\description{
####
Initialize stat.gam.result and chng.gam.result
}
\keyword{internal}
