% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFlow.R
\name{.mergeFlow}
\alias{.mergeFlow}
\title{merge flow variable into analysis data frame and update iSpec with variable name}
\usage{
.mergeFlow(ct1 = ct1, iSpec = iSpec, gageID = gageID,
  hydro.var = hydro.var, flow.detrended = flow.detrended)
}
\arguments{
\item{ct1}{analysis data frame}

\item{iSpec}{iSpec}

\item{gageID}{"q" + USGS gage ID}

\item{hydro.var}{averaging windows}

\item{flow.detrended}{data generated by detrended.flow.  Default = flow.detrended.}
}
\description{
merge flow variable into analysis data frame and update iSpec with variable name
}
\keyword{internal}
