% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSalinity.R
\name{.mergeSalinity}
\alias{.mergeSalinity}
\title{merge salinity into analysis data frame and update iSpec with variable name}
\usage{
.mergeSalinity(ct1 = ct1, iSpec = iSpec,
  salinity.detrended = salinity.detrended)
}
\arguments{
\item{ct1}{analysis data frame}

\item{iSpec}{iSpec}

\item{salinity.detrended}{data generated by detrended.salinity.  Default = detrended.salinity.}
}
\description{
merge salinity into analysis data frame and update iSpec with variable name
}
\keyword{internal}
