% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers2.R
\name{.P}
\alias{.P}
\title{Paragraph  (customization of pandoc.p)}
\usage{
.P(text = NULL)
}
\arguments{
\item{text}{text of paragraph}
}
\value{
n/a
}
\description{
Paragraph  (customization of pandoc.p)
}
\examples{
.P()
}
\seealso{
.F .H .H2 .H3 .H4 .P .T .V
}
