% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeCensored.R
\name{.simCensored}
\alias{.simCensored}
\title{Fit normal or log normal distribution; simulation in censored data
provided by Elgin Perry
phi.left and phi.right are distribution function, e.g., pnorm(q=10,mean=10,sd=1) =0.5
phi.sim is a uniform random number between phi.left and phi.right
sim is imputed value}
\usage{
.simCensored(x, distr = "norm")
}
\arguments{
\item{x}{data}

\item{distr}{Distribution type; norm (default) or lnorm}
}
\value{
qnorm (or qlnorm) values for input data
}
\description{
Fit normal or log normal distribution; simulation in censored data
provided by Elgin Perry
phi.left and phi.right are distribution function, e.g., pnorm(q=10,mean=10,sd=1) =0.5
phi.sim is a uniform random number between phi.left and phi.right
sim is imputed value
}
\keyword{internal}
