% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers2.R
\name{.T}
\alias{.T}
\title{Print out table title (customization of pandoc.emphasis and pandoc.strong )}
\usage{
.T(text, n = NULL, t = "e")
}
\arguments{
\item{text}{text of table title}

\item{n}{table number}

\item{t}{emphasis or stong}
}
\value{
n/a
}
\description{
Print out table title (customization of pandoc.emphasis and pandoc.strong )
}
\examples{
text<-"Hello World!"
.T(text)
.T(text, 4)
.T(text, 4,'e')
.T(text, 4,'s')
}
\seealso{
.F .H .H2 .H3 .H4 .P .T .V
}
\keyword{internal}
