% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers2.R
\name{.V}
\alias{.V}
\title{Print out text  (blended pandoc.emphasis, .verbatim, and .strong)}
\usage{
.V(text = " ", t = "v")
}
\arguments{
\item{text}{text}

\item{t}{emphasis or stong or verbatim}
}
\value{
n/a
}
\description{
Print out text  (blended pandoc.emphasis, .verbatim, and .strong)
}
\examples{
.V("Hello World!",'v')
.V("Hello World!",'e')
.V("Hello World!",'s')
.V("Hello World!")
}
\seealso{
.F .H .H2 .H3 .H4 .P .T .V
}
