% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers2.R
\name{.vTable}
\alias{.vTable}
\title{Print out character vector table in wrapped mode}
\usage{
.vTable(x = c(" "), width = 65)
}
\arguments{
\item{x}{character vector}

\item{width}{wrap width [default=65]}
}
\description{
Print out character vector table in wrapped mode
}
