% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_extractAIC.censReg.R
\name{extractAIC.censReg}
\alias{extractAIC.censReg}
\title{Extract AIC from a Fitted Model}
\usage{
\method{extractAIC}{censReg}(fit, scale = 0, k = 2, correct = FALSE, ...)
}
\arguments{
\item{fit}{an object of class "censReg"---output from \code{censReg}.}

\item{scale}{set scale, ignored in current version.}

\item{k}{the weight of the equivalent degrees of freedom.}

\item{correct}{logical if \code{TRUE} and \code{k} is set to 2, then return
corrected AIC, if \code{FALSE} or \code{k} is not set to 2,then return the 
uncorrected value, either AIC or BIC.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A vector of length 2, giving the equivalent degrees of freedom and
the AIC for \code{fit}.
}
\description{
Computes the Akaike (1974) An Information Criterion for a censored regression 
object: method for "censReg" object.
}
\references{
Akaike, H., 1974, A new look at the statistical model identification:
IEE transactions on Automatic Control, v. 19, no. 6, p. 716--723.
}
\seealso{
\code{\link{censReg}}, \code{\link{extractAIC}}
}
\keyword{regression}
