% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_fitted.censReg.R
\name{fitted.censReg}
\alias{fitted.censReg}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{censReg}(object, suppress.na.action = FALSE,
  type = c("response", "link", "mean"), ...)
}
\arguments{
\item{object}{an object of class "censReg"---output from \code{censReg}}

\item{suppress.na.action}{logical, suppress the effects of the
\code{na.action} in the call to \code{censReg} and return only the fitted
values corresponding to the fitted data.}

\item{type}{the type of fitted value. To get fitted values on the
original scale of the response, use "response." For back-transformed log-normal 
fitted values, use "link." To get the unbiased estimates of the
mean for log-normal data, use "mean." Note that "link" and "mean" do not work
for the common-log distribution.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The fitted values from the regression.
}
\description{
Extracts the fitted values from a censored regression object: method 
for "censReg" object.
}
\seealso{
\code{\link{censReg}}
}
\keyword{regression}
