% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flwAveragePred.R
\name{flwAveragePred}
\alias{flwAveragePred}
\title{Flow Averaged Predictions}
\usage{
flwAveragePred(pdat, pgam, normPct)
}
\arguments{
\item{pdat}{prediction data set}

\item{pgam}{gam model}

\item{normPct}{vector of percentiles to use in weighting}
}
\value{
data frame of flow[salinity] averaged predictions and standard errors
}
\description{
Compute flow[salinity] weighted predictions
}
