% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamPlotDisp.R
\name{gamPlotDisp}
\alias{gamPlotDisp}
\title{Plot censored gam fits vs. time}
\usage{
gamPlotDisp(gamResult = gamResult, analySpec = analySpec,
  fullModel = 2, seasAvgModel = 2, seasonalModel = 2,
  diffType = "regular", obserPlot = TRUE, interventionPlot = TRUE,
  seasAvgPlot = TRUE, seasAvgConfIntPlot = TRUE,
  seasAvgSigPlot = TRUE, fullModelPlot = TRUE, seasModelPlot = TRUE,
  BaseCurrentMeanPlot = TRUE, adjustedPlot = FALSE)
}
\arguments{
\item{gamResult}{output from procedure gamTest}

\item{analySpec}{analytical specifications}

\item{fullModel}{GAM # for displaying full GAM (e.g., 0, 1, 2)}

\item{seasAvgModel}{GAM # for displaying seasonally average GAM}

\item{seasonalModel}{GAM # for displaying seasonal GAM}

\item{diffType}{plot predicted baseline mean ('regular') or adjusted baseline mean ('adjusted')}

\item{obserPlot}{logical field indicating whether to plot observations}

\item{interventionPlot}{logical field indicating whether to plot interventions (e.g., method changes)}

\item{seasAvgPlot}{logical field indicating whether to plot seasonal average GAM}

\item{seasAvgConfIntPlot}{logical field indicating whether to plot confidence interval for seasonal average GAM}

\item{seasAvgSigPlot}{logical field indicating whether to plot significant increasing and decreasing trends for seasonal average GAM}

\item{fullModelPlot}{logical field indicating whether to plot full GAM}

\item{seasModelPlot}{logical field indicating whether to plot seasonal GAM}

\item{BaseCurrentMeanPlot}{logical field indicating whether to plot baseline and current mean}

\item{adjustedPlot}{logical field indicating whether to plot adjusted model}
}
\description{
Plot censored gam fits vs. time
}
