% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_is.na.R
\name{is.na.lcens}
\alias{is.na.lcens}
\alias{is.na.mcens}
\alias{is.na.qw}
\alias{is.na.mcens}
\alias{is.na.qw}
\title{Missing Values}
\usage{
\method{is.na}{lcens}(x)

\method{is.na}{mcens}(x)

\method{is.na}{qw}(x)
}
\arguments{
\item{x}{the object to be tested.}
}
\value{
A logical vector of the same length as its argument \code{x},
containing TRUE for those elements marked NA and FALSE otherwise.
}
\description{
Indicates which elements are missing: methods for "lcens," "mcens," and "qw" data.  From smwrQW package.
}
\examples{

is.na(as.lcens(c(1,3, NA), 2))

}
\keyword{internal}
\keyword{manip}
