% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layerLukup}
\alias{layerLukup}
\title{Layer Lookup}
\format{A data frame with 10 rows and 3 variables:
\describe{
   \item{layers}{layer abbreviation}
   \item{order}{sort order of layers}
   \item{name}{layer name}
}}
\usage{
layerLukup
}
\description{
Layer Lookup
}
\keyword{datasets}
