% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2_smwrQW_lcens-class.R
\docType{class}
\name{lcens-class}
\alias{lcens-class}
\title{Left-censored Data}
\description{
Class "lcens" describes left-censored data.
From smwrQW package.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a 2-column matrix that contains the value and the detection limit
of the left-censored data. Missing values are permitted.}

\item{\code{censor.codes}}{a logcial value, \code{TRUE} indicates a censored value, 
\code{FALSE} indates an uncensored value. Missing values are permitted only for
missing values in \code{.Data}.}

\item{\code{names}}{a character vector of unique names for each element in \code{.Data}.}
}}

\section{Objects from the Class}{
 Objects can be created by calls to
\code{as.lcens}.
}

\examples{

showClass("lcens")

}
\references{
Lorenz, D.L., in preparation
}
\keyword{classes}
