% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelsResid.R
\name{loadModelsResid}
\alias{loadModelsResid}
\title{Load Built-in GAM formulas for calculating residuals}
\usage{
loadModelsResid(gamSelect = "doy_flw_sal")
}
\arguments{
\item{gamSelect}{character vector of models (Current options include 'doy',
'doy_flw_sal', 'doy_flw_sal_int')}
}
\value{
Returns a list with GAM formulas
}
\description{
Load Built-in GAM formulas for calculating residuals
}
