% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSurvDF.R
\name{makeSurvDF}
\alias{makeSurvDF}
\title{Convert data frame to include Surv objects}
\usage{
makeSurvDF(df)
}
\arguments{
\item{df}{name of data frame}
}
\value{
dataframe with Surv fields
}
\description{
Examines/evaluates dataframe for fields with "_lo" and "_hi" suffixes to convert 
to Surv objects.
}
\examples{
\dontrun{
df <- dataCensored[1:20,]
df1 <- unSurvDF(df)
df2 <- makeSurvDF(df1)
}

}
