% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_mcens-class.R
\docType{class}
\name{mcens-class}
\alias{mcens-class}
\title{Mutliply-censored Data}
\description{
Class "mcens" describes right- or multiply-censored data.
From smwrQW package.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a 2-column matrix that contains the lower and upper value for the data.
Missing values are permitted.}

\item{\code{censor.codes}}{an integer vector that indicates censoring; -1 indicates 
left-censored, 0 indicates uncensored, and 1 indicates rigth-censored.
Missing values are permitted only for elements of \code{.Data} that are missing.}

\item{\code{interval}}{a logical value; \code{TRUE} indicates that the value is 
interval censored and \code{FALSE} indicates otherwise.
Missing values are permitted only for elements of \code{.Data} that are missing.}

\item{\code{names}}{a character vector of unique names for each element in \code{.Data}.}
}}

\note{
Note that \code{censor.codes} and \code{interval} are not necessary, but
included to make some manipulations easier.
}
\section{Objects from the Class}{
 Objects can be created by calls to
\code{as.mcens}.
}

\examples{

showClass("mcens")

}
\references{
Lorenz, D.L., in preparation
}
\keyword{classes}
