% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_mdlAMLE.R
\name{mdlAMLE}
\alias{mdlAMLE}
\title{Estimate Statistics}
\usage{
mdlAMLE(x, method = "AMLE", alpha = 0.4)
}
\arguments{
\item{x}{the data to estimate, Missing values permitted and ignored.
Must be an object of class "lcens," a numeric vector, or the output from censpp.}

\item{method}{the method to use, either "AMLE" or "log AMLE."}

\item{alpha}{the offset for plotting postion, used to compute the filled in values.}
}
\value{
A list containing the mean and standard deviation, filled in
values for the censored values, and the maximum censor level. If \code{method}
is "log AMLE," then the list also contains the mean and standard deviation of the 
natural log-transformed values computed by maximum likelihood.
}
\description{
Support function for computing statistics for left-censored data using the
adjusted maximum-likelihood method (Cohn, 1988)
}
\references{
Cohn, T.A., 1988, Adjusted maximum likelihood estimation of the moments
of lognormal populations from type 1 censored samples: U.S. Geological Survey 
Open-File Report 88-350, 34 p.
}
\keyword{internal}
\keyword{misc}
