% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_mdlMLE.R
\name{mdlMLE}
\alias{mdlMLE}
\title{Estimate Statistics}
\usage{
mdlMLE(x, method = "MLE", alpha = 0.4)
}
\arguments{
\item{x}{the data to estimate, Missing values permitted and ignored.
Must be an object of class "lcens," a numeric vector, or the output from censpp.}

\item{method}{the method to use, either "MLE" or "log MLE."}

\item{alpha}{the offset for plotting position, used to compute the filled in values.}
}
\value{
A list containing the mean and standard deviation, filled in
values for the censored values, and the censored levels. If \code{method}
is "log MLE," then the list also contains the mean and standard deviation of the 
natural log-transformed values computed by maximum likelihood.
}
\description{
Support function for computing statistics for left-censored data using the 
maximum likelihood method (Helsel and Cohn, 1988).
}
\references{
Helsel, D.R. and Cohn, T.A., 1988, Estimation of descriptive statistics 
for multiply censored water quality data: Water Resources Research v. 24, n.
12, p.1997--2004
}
\keyword{internal}
\keyword{misc}
