% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_meanStats.R
\name{mean.lcens}
\alias{mean.lcens}
\alias{mean.mcens}
\alias{mean.mcens}
\title{Arithmetic Mean}
\usage{
\method{mean}{lcens}(x, ...)

\method{mean}{mcens}(x, ...)
}
\arguments{
\item{x}{the censored data object.}

\item{\dots}{not used, require for other methods.}
}
\value{
An error message.
}
\description{
Computes the arithemtic mean. These functions return \code{NA} and are 
designed to stop the user from accidentally using the function \code{mean} 
on censored data.
}
\note{
The mean and standard deviation of censored data should be computed
using \code{censStats}.
}
\seealso{
\code{\link{censStats}}
}
\keyword{censored}
\keyword{internal}
\keyword{univariate}
