% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_pairedPratt.test.R
\name{pairedPratt.test}
\alias{pairedPratt.test}
\title{Test for difference in left-censored samples}
\usage{
pairedPratt.test(x, y, alternative = "two.sided", mu = 0, data.names)
}
\arguments{
\item{x}{the paired samples to \code{y}. Forced to class "lcens." Missing 
values are removed before the analysis.}

\item{y}{the paired values to \code{x}. Missing values are removed before
the analysis.}

\item{alternative}{character string describing the alternative hypothesis for
(\code{x} - \code{y} compared to \code{mu}. Must be one of "two.sided," 
"greater," or "less."}

\item{mu}{a numeric value specifying the test difference between \code{x}
and \code{y}.}

\item{data.names}{character string to be used to explain the data. Default
names are derived from the data arguments.}
}
\value{
An object of class "htest" that inherits "ppw."
}
\description{
Tests for differences in paired left-censored samples from two
groups using Pratt's (1959) adjustment for ties.
}
\note{
The Pratt (1959) adjustment to the Wilcoxon signed-rank test was
used by Lindsey and Rupert (2012) to evaluate decadal changes in groundwater 
quality. Lindsey and Rupert (2012) used a fixed reporting level (0.06,
the largest among all of the data) and used a simple
substitute value 0.0599 for the censored values. The approach in \code{pairedPratt.test}
is to recensor values at the largest reporting limit, use one-half that value as the 
simple substitute and round differences greater than the simple substitute value
to multiples of the largest reporting limit; nonzero differences less than or equal to the
simple substitute value are rounded to that value. That rounding scheme
eliminates the ambiguity of comparing interval values created by the difference
between a censored value and an uncensored value.

A \code{plot} method is supported for the returned object.
}
\section{Null Hypothesis}{
 The null hypothesis is that the
distributions are not different from one another.
}

\examples{

# Compare uncensored results
set.seed(699)
Xu <- sort(rlnorm(22, 0, 1))
Yu <- sort(rlnorm(22, .425, 1))
# Treat as paired samples
pairedPratt.test(Xu, Yu)
wilcox.test(Xu, Yu, paired=TRUE)
# The differences in the p-values are due to use of approximate p-value for
# the Pratt test
}
\references{
Lindsey, B.D., and Rupert, M.G., 2012, Methods for evaluating temporal 
groundwater quality data and results of decadal-scale changes in chloride, 
dissolved solids, and nitrate concentrations in groundwater in the United States, 
1988--2010: U.S. Geological Survey Scientific Investigations Report 2012--5049, 46 p.

Pratt, J.W., 1959, Remarks on zeros and ties in the Wilcoxon signed rank 
procedures: Journal of the American Statistical Association, v. 54, no. 287
p. 655--667.
}
\seealso{
\code{\link{ppw.test}}, \code{\link{lcens-class}}
}
\keyword{censored}
\keyword{htest}
