% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_pcodeNWISqw.R
\name{pcodeNWISqw}
\alias{pcodeNWISqw}
\title{Parameter Code Information}
\usage{
pcodeNWISqw(params = "all", group = TRUE, name = TRUE, CASRN = FALSE,
  short = TRUE, units = TRUE, col.name = FALSE)
}
\arguments{
\item{params}{A character string contains the name of a group of parameter
codes, or a vector of pamater codes. See \bold{Details}.}

\item{group}{include the parameter group in the output?}

\item{name}{include the parameter name in the output?}

\item{CASRN}{include the parameter Chemical Abstracts Service Registry Number
(CASRN) in the output?}

\item{short}{include the parameter short name in the output?}

\item{units}{include the parameter units in the output?}

\item{col.name}{include a column name for the parameter in the output?}
}
\value{
A data frame with the parameter codes and the selected columns.
}
\description{
Returns information about parameter codes. Support function for \code{importNWISqw}.
}
\details{
Valid groups for \code{param} are "All," "INF," "PHY," "INM," "INN," 
"NUT," "MBI," "BIO," "IMN," "IMM," "TOX," "OPE," "OPC," "OOT," "RAD,"
"SED," or "POP."
}
\examples{

\dontrun{
pcodeNWISqw("00925")
}

}
\references{
Lorenz, D.L., 2014, smwrQW OFR.\cr See information about discrete
samples at \url{https://nwis.waterdata.usgs.gov/usa/nwis/qw}.
}
\seealso{
\code{\link[dataRetrieval]{readNWISpCode}}
}
\keyword{IO}
\keyword{datasets}
