% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_pctCens.R
\name{pctCens}
\alias{pctCens}
\title{Percent Censored}
\usage{
pctCens(x, type = "multiple")
}
\arguments{
\item{x}{the object to compute the percentage of censoring. Missing values
are permitted and removed before computing the percentage,}

\item{type}{the type of censoring desired. For "left," only the percent of
left-censored data is computed. For "right," only the percent of
right-censored data is computed. For "multiple," the percentage of left-, right-,
and interval-censored data is returned.}
}
\value{
The percentage of the kind of censoring specified by \code{type} is
returned. Zero is returned if all data are missing!
}
\description{
Computes the percentage of censoring ("left," "right," or "multiple") for an 
object.
}
\note{
This function will work for any type of data that has a method for 
\code{censoring} and \code{as.mcens}.
}
\examples{

pctCens(as.lcens(c(1:5, NA), 2)) # left-censored 

}
\seealso{
\code{\link{censoring}}
}
\keyword{attribute}
\keyword{censored}
