% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_pow.R
\name{pow}
\alias{pow}
\title{Power Transforms}
\usage{
pow(x, lambda, out = "Auto")
}
\arguments{
\item{x}{any object that can be converted to class "mcens" or "lcens." Must be non negative (zero data are
permitted). Missing values are permitted and missing vlues in the returnned vector.}

\item{lambda}{the power to raise \code{x}. Must be strictly positive.}

\item{out}{the output format, must be one of "numeric," "lcens," "mcens," or "Auto." If "Auto," then if
\code{x} is class "numeric," then the output format is "numeric" otherwise the output format is "mcens."}
}
\value{
An object of class "mcens" that is the result of the requested operations. Each of the range of the 
lower and upper values are raised to the power of \code{lambda} and the result divided by \code{lambda}
to approximately maintain scale. But the reciprocal value of \code{lambda} cannot be used to restore the 
original values.
}
\description{
Raises censored data to a positive power.
}
\examples{
pow(as.mcens(c(0,2), c(1,2)), 2)
pow(as.mcens(c(0,2), c(1,2)), .5)
# Numeric values
pow(c(1,3), 2)
}
