% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_ppw.test.R
\name{ppw.test}
\alias{ppw.test}
\title{Test for difference in left-censored samples}
\usage{
ppw.test(x, y, alternative = "two.sided", OBrienFleming = TRUE, data.names)
}
\arguments{
\item{x}{the paired samples to \code{y}. Forced to class "lcens." Missing 
values are removed before the analysis.}

\item{y}{the paired values to \code{x}. Missing values are removed before
the analysis.}

\item{alternative}{character string describing the alternative hypothesis.
Must be one of "two.sided," "greater," or "less."}

\item{OBrienFleming}{logical, if \code{TRUE}, then apply the OBrienFleming otherwise it is
not applied. See \bold{Details}.}

\item{data.names}{character string to be used to explain the data. Default
names are derived from the data arguments.}
}
\value{
An object of class "htest" that inherits "ppw."
}
\description{
Tests for differences in paired left-censored samples from two
groups using the PPW test (O'Brien and Fleming, 1987).
}
\details{
The O'Brien-Fleming adjustment forces the score for all paired censored
values to be equal to each other. It also forces equality between observed
values that are less than the detection limit in the paried sample by setting
the value to less than the detection limit.
}
\note{
The \code{ppw.test}
function uses the \code{survfit} function. Helsel (2012) describes flipping
the left-censored data so that small values become large and left-censored
values become right-censored values and adapt nonparametric techniques from
survival analysis.\cr
A \code{plot} method is supported for the returned object.
}
\section{Null Hypothesis}{
 The null hypothesis is that the
distributions are not different from one another.
}

\examples{

# Compare uncensored results
set.seed(699)
Xu <- sort(rlnorm(22, 0, 1))
Yu <- sort(rlnorm(22, .425, 1))
# Treat as paired samples
ppw.test(Xu, Yu)
wilcox.test(Xu, Yu, paired=TRUE)

}
\references{
Helsel, D.R. 2012, Statistics for Censored Environmental Data Using Minitab
and R: New York, Wiley, 324 p.\cr

O'Brien, P.C. and Fleming, T.R., 1987, A paired Prentice-Wilcoxon test for censored 
paried data: Biometrics, v. 43, p. 451--455.
}
\seealso{
\code{\link{survdiff}}, \code{\link{survfit}},
\code{\link{lcens-class}}
}
\keyword{censored}
\keyword{htest}
\keyword{internal}
