% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_print.censMCT.R
\name{print.censMCT}
\alias{print.censMCT}
\title{Print Results}
\usage{
\method{print}{censMCT}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class "censMCT"---output from \code{censMulticomp.test}}

\item{digits}{the number of significant digits to print}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Print the results of mutliple nonparametric comparison tests among groups 
of left-censored data.
}
\note{
The printed output includes the name of the test, the error rate, the
method for adjusting the p-values for the individual tests, the names of the
response and group variables, and a table of the paired comparisons. The 
table includes the names of the groups comapred, the z-score from the
\code{genWilcox.test}, the adjusted p-value, and a column to indicate if the
adjusted p-value is less than the \code{alpha} level set in
\code{censMulticomp.test}.
}
\seealso{
\code{\link{censKSample.test}}, \code{\link{censMulticomp.test}},
\code{\link{genWilcox.test}}
}
\keyword{utilities}
