% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_print.censQuantile.R
\name{print.censQuantile}
\alias{print.censQuantile}
\title{Print Results}
\usage{
\method{print}{censQuantile}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class "censQuantile."}

\item{digits}{the number of significant digits to print}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Print the output of \code{censQuantile}.
}
\note{
The printed output appears as 2 columns, the first column is the
percentile and the second column is the value.
}
\keyword{utilities}
