% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_print.censReg.R
\name{print.censReg}
\alias{print.censReg}
\title{Print Results}
\usage{
\method{print}{censReg}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class "censReg"---output from \code{censReg}}

\item{digits}{the number of significant digits to print}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Print the results of an AMLE/MLE regression.
}
\note{
The printed output includes the call, the coefficent table, the
estimated residual standard error, the log-likelihood of the model and
null model with the attained p-value, and the computational method.
}
\seealso{
\code{\link{censReg}}, \code{\link{coef.censReg}}
}
\keyword{utilities}
