% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Q05078470}
\alias{Q05078470}
\title{Daily Streamflow}
\format{
Data frame with 365 rows and 2 columns\cr
\tabular{lll}{ 
Name \tab  Type \tab Description\cr\cr
DATES \tab Date \tab Day\cr
FLOW \tab numeric \tab Daily mean streamflow\cr
} Note: the daily mean streamflow rates are in
cubic feet per second.
}
\source{
Data retrieved from NWISWeb
(\url{https://nwis.waterdata.usgs.gov/mn/nwis/sw}) on 2005-01-05.
}
\usage{
Q05078470
}
\description{
Daily mean flow for Judicial Ditch 64 near Mentor, Minnesota (SW4), USGS
station number 05078470 for calendar year 2003.

From smwrData.
}
\examples{
data(Q05078470)
# Plot the data for station 05078470
with(Q05078470, plot(DATES, FLOW, type='l'))
}
\keyword{datasets}
