% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qw.export.R
\name{qw.export}
\alias{qw.export}
\title{qw.export}
\usage{
qw.export(df.qw, dir.output = getwd(), output.fn.prefix = "qw")
}
\arguments{
\item{df.qw}{Data frame with qw fields to be exported.}

\item{dir.output}{Directory where output files will be saved. 
Default is working directory.}

\item{output.fn.prefix}{Prefix for output files.
Files will be named with prefix and QW column name.  Default is "qw".}
}
\value{
Returns a series of tab delimited files to the user defined directory 
matching the number of qw fields in the input data frame.
}
\description{
Export the contents of a QW object to tab delimited text files.
}
\details{
This internal function will export each qw field of a given data frame 
to a separate tab separated file (tsv).  The left header information wil be attached to each file.
The following columns will be named with the qw column named and the qw slot names (e.g., cond.remark.codes). 
The exported files will be in a directory specified by the user (defaults to working directory)
and the files will be named with the name of the qw columns with a user 
defined prefix (defaults to qw.)
}
\examples{
# define data frame with qw column classes
myDF <- dataCensored

# directory to save output
dir.save <- getwd()

# prefix for file names ()
fn.prefix <- "dataCensored" 

# run function
qw.export(myDF, dir.save, fn.prefix)
}
\keyword{internal}
