% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_qw2mcens.R
\name{qw2mcens}
\alias{qw2mcens}
\title{Conversion to class "mcens"}
\usage{
qw2mcens(x)
}
\arguments{
\item{x}{an object of class "qw" to convert.}
}
\value{
A vector of the data converted to class "mcens" retaining interval 
censoring for "less-than" data.
}
\description{
Converts objects of class "qw" to "mcens" retaining interval censoring
for "less-than" data.
}
\seealso{
\code{\link{as.mcens}}
}
