% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_readNWQLdl.R
\name{readNWQLdl}
\alias{readNWQLdl}
\title{Historical Reporting Limits}
\usage{
readNWQLdl(parm_cd)
}
\arguments{
\item{parm_cd}{the parameter code.}
}
\value{
A data frame of the analyte, methods, begin and end date for each
reporting level change, the reporting level type used, the long-term
detection limit and the reporting level.
}
\description{
Imports the historical reporting limits for all methods for a 
parameter code.
}
\note{
This function works only within the internal USGS network.
}
\examples{

\dontrun{
# This function works only within the internal USGS network.
#readNWQLdl("00608")
}

}
\references{
Lorenz, D.L., 2014, USGSqw OFR.
}
\seealso{
\code{\link{qw-class}}
}
\keyword{IO}
\keyword{internal}
