% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sal}
\alias{sal}
\title{Salinity data}
\format{A data frame with 51,092 rows and 4 variables:
\describe{
   \item{Station}{CBP Station ID}
   \item{Sample_Date}{date, YYYY-MM-DD}
   \item{Layer}{sample layer}
   \item{salinity}{Measured salinity}
}}
\usage{
sal
}
\description{
Salinity data, 1984 to 2016, for 8 stations.
}
\keyword{datasets}
