% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_sdlFill.R
\name{sdlFill}
\alias{sdlFill}
\title{Estimate Statistics}
\usage{
sdlFill(x, method = "fill", alpha = 0.4)
}
\arguments{
\item{x}{the data to estimate, Missing values permitted and ignored.
Must be an object of class "lcens," a numeric vector, or the output from censpp.}

\item{method}{the method to use, either "fill" or "log fill"}

\item{alpha}{the offset for plotting postion (not used, required for other methods).}
}
\value{
A list containing the mean and standard deviation, filled in
values for the censored values, and the maximum censor level.
}
\description{
Support function for computing statistics for left-censored data.
}
\note{
This function uses the method of Gleit (1985). That method computes
the mean and standard deviation of the data, estimates synthetic values for
the censored values using a method similar to ROS, and repeats the process until the 
mean and standard deviation stabilize.
}
\references{
Gleit, A., 1985, Estimation for small normal data sets with detection limits:
Environmental Science and Technology, v. 19, no. 12, p. 1201--1206.
}
\keyword{misc}
