% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrGraphs_strip.blanks.R
\name{strip.blanks}
\alias{strip.blanks}
\title{Remove Spaces}
\usage{
strip.blanks(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
A vector like \code{x}, but with leading and trailing spaces removed
from each element.
}
\description{
Removes leading and trailing blanks from a character string.  From smwrGraphs package.
}
\examples{

strip.blanks("   keep me   ")

}
\seealso{
\code{link[base]{sub}}
}
\keyword{internal}
\keyword{manip}
