\name{usgsGages}
\alias{usgsGages}
\docType{data}
\title{
Table: Core USGS gages for CBP trend analyses
}
\description{
Core USGS gages for CBP trend analyses
}
\usage{data("usgsGages")}
\format{
  A data frame with the following 2 variables.
  \describe{
    \item{usgsGageID}{USGS site ID (i.e., gage number)}
    \item{\code{siteName}}{USGS site location}
  }
}
\examples{
# Show sites core USGS gages
str(usgsGages)
}
\keyword{datasets}
