% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_vcov.censReg.R
\name{vcov.censReg}
\alias{vcov.censReg}
\title{Variance-Covariance Matrix}
\usage{
\method{vcov}{censReg}(object, ...)
}
\arguments{
\item{object}{an object of class "censReg"---output from \code{censReg}.}

\item{\dots}{not used, further arguments passed to or from other methods.}
}
\value{
A matrix of the estimated covariances between the parameter 
estimates for each main parameter of the model. The matrix has row and 
column names corresponding to the parameter names.
}
\description{
Extracts the variance-covariance matrix of the main parameters from a 
censored regression object: method for "censReg" object.
}
\seealso{
\code{\link{censReg}}
}
\keyword{internal}
\keyword{regression}
