% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smwrQW_x_z%eq%.R
\name{\%~=\%}
\alias{\%~=\%}
\title{Possible Equality}
\usage{
e1 \%~=\% e2
}
\arguments{
\item{e1}{any vector that can be converted to class "mcens."}

\item{e2}{any vector that can be converted to class "mcens."}
}
\value{
A logical vector that matches the length of the longer of \code{e1} or 
\code{e2}, set to \code{TRUE} is there is overlap in the values, and \code{FALSE}
otherwise.
}
\description{
Checks for the possibility that two values are equal. If there is any overlap
in range, they are considered possibly equal and \code{TRUE} is returned, otherwise
\code{FALSE} is returned.
}
\details{
Missing values are permitted in either \code{e1} or \code{e2}. wherever either 
are missing the returned element is set to \code{NA}.
}
\examples{
as.mcens(1, 2) \%~=\% 0.5
as.mcens(1, 2) \%~=\% 1.5
as.mcens(1, 2) \%~=\% 2.5
}
\seealso{
\code{\link{Compare,qw,numeric-method}}
}
\keyword{internal}
\keyword{manip}
