% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baytrends.R
\docType{package}
\name{baytrends}
\alias{baytrends}
\alias{baytrends-package}
\title{baytrends: Long Term Water Quality Trend Analysis}
\description{
The baytrends package was developed to enable users to evaluate long-term
trends in the Chesapeake Bay using a Generalized Additive Modeling (GAM)
approach. This package includes features that allow for explicit modeling
of left- and interval-censored data as well as monitoring data that 
includes laboratory method changes.
}
\section{baytrends functions}{

analysisOrganizeData
as.qw
as.qw-method
...
}

