% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baytrends.R
\docType{package}
\name{baytrends}
\alias{baytrends}
\alias{baytrends-package}
\title{baytrends: Long Term Water Quality Trend Analysis}
\description{
The baytrends package was developed to enable users to evaluate long-term
trends in the Chesapeake Bay using a Generalized Additive Modeling (GAM)
approach. The model development includes selecting a GAM structure to
describe nonlinear seasonally-varying changes over time, incorporation of
hydrologic variability via either a river flow or salinity, the use of an
intervention to deal with method or laboratory changes suspected to impact
data values, and representation of left- and interval-censored data. This
approach, which is fully transferable to other systems, allows for Chesapeake
Bay water quality data to be evaluated in a statistically rigorous, yet
flexible way to provide insights to a range of management- and
research-focused questions.
}
