% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataCensored}
\alias{dataCensored}
\title{Censored Data Example}
\format{A data frame with 7719 rows and 16 variables:
\describe{
   \item{station}{Station code}
   \item{date}{}
   \item{layer}{}
   \item{secchi}{}
   \item{chla}{}
   \item{do}{}
   \item{tn}{}
   \item{tp}{}
   \item{po4f}{}
   \item{pp}{}
   \item{tdp}{}
   \item{no23f}{}
   \item{nh4f}{}
   \item{tdn}{}
   \item{pn}{}
   \item{tss}{}
}}
\source{
Chesapeake Bay data
}
\usage{
dataCensored
}
\description{
A dataset with example censored data.
}
\keyword{datasets}
