% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataCensored}
\alias{dataCensored}
\title{Chesapeake Bay Program Monitoring Data, 1985-2016}
\format{
A data frame with 13,062 rows and 17 variables:
\describe{
  \item{station}{station identifier}
  \item{date}{sample date}
  \item{layer}{sample layer}
  \item{secchi}{Secchi depth [m]}
  \item{salinity}{Salinity [ppt]}
  \item{do}{Dissolved Oxygen [mg/L]}
  \item{wtemp}{Water Temperature [deg C]}
  \item{chla}{Corrected Chlorophyll-a [ug/L]}
  \item{tn}{Total Nitrogen [mg/L]}
  \item{tp}{Total Phosphorus [mg/L]}
  \item{tss}{Total Suspended Solids [mg/L]}
  \item{din}{Dissolved Inorganic Nitrogen [mg/L]}
  \item{po4}{Orthophosphorus [mg/L]}
  \item{tdn}{Total Dissolved Nitrogen [mg/L]}
  \item{tdp}{Total Dissolved Phosphorus [mg/L]}
  \item{nh4}{Ammonium [mg/L]}
  \item{no23}{Nitrite + Nitrate [mg/L]}
}
}
\usage{
dataCensored
}
\description{
Selected 1985-2016 data for eight (8) stations from the
  Chesapeake Bay Monitoring Program. Water quality variables are stored as
  either class 'num' for variables with no censoring or class 'Surv' 
  (see survival::Surv) that allows for left- and interval-censored data.
}
\keyword{datasets}
