% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrended.flow.R
\name{detrended.flow}
\alias{detrended.flow}
\title{Detrended Flow Data}
\usage{
detrended.flow(usgsGageID, siteName, yearStart, yearEnd, dvAvgWinSel,
  dvAvgWgtSel, dvAvgSidesSel, lowess.f, dir.out, fn.out)
}
\arguments{
\item{usgsGageID}{USGS GageIDs (e.g., 01491000)}

\item{siteName}{USGS SiteName (not used for matching only for plot titles)}

\item{yearStart}{USGS S}

\item{yearEnd}{USGS S}

\item{dvAvgWinSel}{USGS S}

\item{dvAvgWgtSel}{USGS S}

\item{dvAvgSidesSel}{USGS S}

\item{lowess.f}{USGS S}

\item{dir.out}{Output directory for data file, plots, and such}

\item{fn.out}{data file output name (will save as RDA).}
}
\value{
Creates an RDA file in the user specified directory 
                  and creates plots (but does not save them).
}
\description{
Create data set, detrend it, create stats, and plot it.  For flow and salinity.
}
\details{
Input is a list of USGS gages and other
}
\examples{

# Define Function Inputs
usgsGageID    <- c("01491000", "01578310")
siteName      <- c("Choptank River near Greensboro, MD",
                   "Susquehanna River at Conowingo, MD")
yearStart     <- 1983
yearEnd       <- 2016
dvAvgWinSel   <- c(1, 5, 10, 15, 20, 30, 40, 50, 60, 90, 120, 150, 180, 210)
dvAvgWgtSel   <- "uniform"
dvAvgSidesSel <- 1
lowess.f      <- 0.2
dir.out       <- file.path(getwd(), "data2")
fn.out        <- paste(yearStart, yearEnd
                       ,"seasonally_detrended_flow_data.rda", sep="_")
                 
# Run Function
detrended.flow(usgsGageID, siteName, yearStart, yearEnd
              , dvAvgWinSel, dvAvgWgtSel, dvAvgSidesSel
              , lowess.f, dir.out, fn.out)
              
}
