% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamTables.R
\name{.gamCoeff}
\alias{.gamCoeff}
\title{Prepare table of coefficients for GAM analysis}
\usage{
.gamCoeff(lmo, iSpec)
}
\arguments{
\item{lmo}{output from gam model}

\item{iSpec}{data frame with intervenList}
}
\description{
Prepare table of coefficients for GAM analysis
}
\keyword{internal}
