% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow.detrended}
\alias{flow.detrended}
\title{Detrended Flow Data}
\format{A list of 8:
\describe{
   \item{retreiveDate}{Data retrieval date}
   \item{gages}{data frame with 2 rows and 2 variables: usgsGageID and siteName}
   \item{yearStart}{Start year}
   \item{yearEnd}{End year}
   \item{dvAvgWinSel}{}
   \item{dvAvgWgtSel}{}
   \item{dvAvgSidesSel}{}
   \item{lowess.f}{}
}}
\usage{
flow.detrended
}
\description{
Detrended Flow Data
}
\keyword{datasets}
