% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow.detrended}
\alias{flow.detrended}
\title{Detrended flow data}
\format{A list of 8:
\describe{
   \item{retreiveDate}{Date data was generated}
   \item{gages}{data.frame: 2 obs of 2 variables}
   \item{usgsGageID}{USGS Gage ID}
   \item{siteName}{USGS Site Name}
   \item{yearStart}{start year, YYYY-MM-DD}
   \item{yearEnd}{end year, YYYY-MM-DD}
   \item{dvAvgWinSel}{daily value, average win sel}
   \item{dvAvgWgtSel}{daily value, average weighted sel}
   \item{dvAvgSidesSel}{daily value, average sides sel} 
   \item{lowess.f}{lowess.f}  
}}
\usage{
flow.detrended
}
\description{
Flow data, 1983 to 2016, seasonally detrended, for 2 stations.
}
\keyword{datasets}
