% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamTest.r
\name{gamTest}
\alias{gamTest}
\title{Perform GAM analysis}
\usage{
gamTest(
  df,
  dep,
  stat,
  layer = NA,
  analySpec,
  gamTable = TRUE,
  gamPlot = 10,
  gamDiffModel = NA,
  flow.detrended = NA,
  salinity.detrended = NA
)
}
\arguments{
\item{df}{data frame}

\item{dep}{dependent variable}

\item{stat}{station}

\item{layer}{layer}

\item{analySpec}{analytical specifications}

\item{gamTable}{gam table setting (set to \code{FALSE} to turn off table output)}

\item{gamPlot}{gam plot setting (set to \code{FALSE} to turn off plotting)}

\item{gamDiffModel}{GAM model(s) used for computing differences on
sub-annual/multi-period basis}

\item{flow.detrended}{data generated by \code{detrended.flow} and must be
specified to run flow adjustment models.}

\item{salinity.detrended}{data generated by \code{detrended.salinity} and
must be specified to run salinity adjustment models.}
}
\value{
Returns a list with the following elements:

\itemize{
\item \strong{stat.gam.result --} Data frame with summary results in a tabular format suitable for output to a csv file
\item \strong{chng.gam.result --} Data frame with estimate of change for sub-annual differences based on the 
settings specified in \code{gamDiffModel}. Includes all data from stat.gam.result and sub-annual estimates of change
in a tabular format suitable for output to a csv file. 
\item \strong{data --} Data frame of data used in analysis (for example years with large proportion of censored data might be removed from the analysis and not included in this data frame)
\item \strong{data.all --} Data frame of all potential data available for analysis
\item \strong{iSpec --} List of baytrends analysis settings used for a specific analysis of station and parameter
\item \strong{gamOutput0 --}	Results including model predictions for gam formula 0 (i.e., gam0)
\item \strong{gamOutput1 --}	Results including model predictions for gam formula 1 (i.e., gam1)
\item \strong{gamOutput2 --}	Results including model predictions for gam formula 2 (i.e., gam2)
\item \strong{gamOutput3 --}	Results including model predictions for gam formula 3 (i.e., gam3)
\item \strong{gamOutput4 --}	Results including model predictions for gam formula 4 (i.e., gam4)
\item \strong{gamOutput5 --}	Results including model predictions for gam formula 5 (i.e., gam5)
\item \strong{gamOutput6 --}	Results including model predictions for gam formula 6 (i.e., gam6)
}

\strong{gamOutput* --} For each evaluated model, \code{gamOutput*} (see above
element list) is a list with the following elements:

\itemize{
\item \strong{gamOption --} gam formula ID, i.e., 0, 1, 2, 3, 4 , 5, 6 corresponding to gam0, gam1, gam2, etc.
\item \strong{gamRslt --} mgcv::gam output
\item \strong{gamRsltSum --} "summary" of mgcv::gam, i.e., summary(mgcv::gam)
\item \strong{gamANOVAtbl --} GAM Analysis of Variance table 
\item \strong{gamCoefftbl --} GAM Parameter Coefficients table
\item \strong{gamDiagnostics --} GAM Diagnostics table(AIC, RMSE, and Adj. R-squared)
\item \strong{perChange --} Estimates of Change table
\item \strong{porDiff.regular --} Estimate of change in list format
\item \strong{porDiff.adjusted --} Estimate of change adjusted for interventions in list format 
\item \strong{predictions --} Data frame of gam predictions (all information used to create baytrends graphics excluding measurements)
}
}
\description{

Perform GAM analysis. Relies on mgcv::gam to perform general additive model.
\code{\link[mgcv]{gam}}
The baseDay function has been added to this package from the smwrBase package.

}
\examples{
# Specify parameter and station to analyze
dep        <- 'do'
stat       <- 'CB5.4'
layer      <- 'B'

# Prepare data and set up specifications for analysis
dfr <- analysisOrganizeData (dataCensored)
df        <- dfr[[1]]
analySpec <- dfr[[2]]

# Apply gamTest 
gamResult <- gamTest(df, dep, stat, layer, analySpec=analySpec)


}
