\name{layerLukup}
\alias{layerLukup}
\docType{data}
\title{
Table: Layer lookup list
}
\description{
Layer lookup list.
}
\usage{data("layerLukup")}
\format{
  A data frame with the following 3 variables.
  \describe{
    \item{\code{layers}}{Layer codes}
    \item{\code{order}}{Analysis order}
    \item{\code{name}}{Layer name}
  }
}
\examples{
data(layerLukup)
}
\keyword{datasets}
