% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layerLukup}
\alias{layerLukup}
\title{Layer List}
\format{A data frame with 10 rows and 3 variables:
\describe{
   \item{layers}{Layer codes}
   \item{order}{Analysis order}
   \item{name}{Layer name}
}}
\usage{
layerLukup
}
\description{
A lookup table of layer abbreviations and the corresponding
  names.
}
\keyword{datasets}
