% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModels.R
\name{loadModels}
\alias{loadModels}
\title{Load Built-in GAM formulas}
\usage{
loadModels(gamSelect = "gam4")
}
\arguments{
\item{gamSelect}{character vector of models (Current options include gam0,
gam1, gam2, gam3, gam4, gam5)}
}
\value{
Returns a list with GAM formulas
}
\description{
Returns built-in GAM formulas
}
