% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSurvDF.R
\name{makeSurvDF}
\alias{makeSurvDF}
\title{Convert dataframe to include survival (Surv) objects}
\usage{
makeSurvDF(df)
}
\arguments{
\item{df}{name of data frame}
}
\value{
dataframe with Surv fields
}
\description{
Within a dataframe, numeric fields using a "_lo" and "_hi" suffix 
naming convention (e.g., "conc_lo" "conc_hi") are combined into a single Surv object
(e.g., "conc") using the "interval2" option provided by through the
\code{survival::Surv(conc_lo, conc_hi, type = "interval2")} syntax.
}
\details{
Converting fields to Surv objects works with field pairs that have a "_lo" 
and "_hi" suffix naming convention. The numeric value for "_hi" must be greater than 
or equal to the "_lo" value. Source data that violate this requirement are 
set to NA with a summary report outputted to the console.
}
\examples{
\dontrun{
df <- dataCensored[1:20,]
df1 <- unSurvDF(df)
df2 <- makeSurvDF(df1)
}

}
