% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameterList}
\alias{parameterList}
\title{Parameter List}
\format{A data frame with 81 rows and 13 variables:
\describe{
   \item{parm}{Parameter}
   \item{parmSource}{Original Parameter}
   \item{parmCat}{Category}
   \item{parmName}{Name}
   \item{parmCalc}{Calculated?}
   \item{parmNamelc}{}
   \item{parmUnits}{Units}
   \item{parmRecensor}{}
   \item{parmRO1}{}
   \item{parmRO2}{}
   \item{parmTrend}{}
   \item{logTrans}{}
   \item{trendIncrease}{}
}}
\usage{
parameterList
}
\description{
Parameter List
}
\keyword{datasets}
