% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameterList}
\alias{parameterList}
\title{Parameter List}
\format{A data frame with 81 rows and 13 variables:
\describe{
   \item{parm}{}
   \item{parmSource}{}
   \item{parmCat}{}
   \item{parmName}{}
   \item{parmCalc}{}
   \item{parmNamelc}{}
   \item{parmUnits}{}
   \item{parmSource}{}
   \item{parmRecensor}{}
   \item{parmRO1}{}
   \item{parmRO2}{}
   \item{parmTrend}{}
   \item{logTrans}{}
   \item{order}{}
   \item{trendIncrease}{}
}}
\usage{
parameterList
}
\description{
Long Description
}
\keyword{datasets}
