% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stationMasterList}
\alias{stationMasterList}
\title{Master Stations List}
\format{A data frame with 145 rows and 19 variables:
\describe{
   \item{station}{}
   \item{state}{}
   \item{locationType}{}
   \item{waterbody}{}
   \item{latitude}{}
   \item{longitude}{}
   \item{cbSeg92}{}
   \item{usgsGageID}{}
   \item{usgsGageMatch}{}
   \item{stationRO1}{}
   \item{stationRO2}{}
   \item{stationGrpName}{}
   \item{stationMethodGroup}{}
   \item{hydroTerm}{}
   \item{flwAvgWin}{}
   \item{flwParms}{}
   \item{salParms}{}
   \item{notes}{}
}}
\usage{
stationMasterList
}
\description{
Long Description
}
\keyword{datasets}
