% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unSurv.R
\name{unSurv}
\alias{unSurv}
\title{Converts Surv object into a 3-column matrix}
\usage{
unSurv(x, col_lo = "lo", col_hi = "hi")
}
\arguments{
\item{x}{vector (Surv object)}

\item{col_lo}{Output column name for "lo" values.  Default = "lo"}

\item{col_hi}{Output column name for "hi" values.  Default = "hi"}
}
\value{
Returns a 3-column matrix: lo, hi, type
}
\description{
Converts Surv object into a 3-column matrix
}
\details{
The type column has the following meanings:

 1  -- no censoring
 
 2  -- left censored ("less than in a survival sense" [-Inf to 10], i.e., <10)
 
 3  -- interval censored ("less than in a water quality sense", i.e., "0 - <3", "1 - 3")
 
 NA -- missing value
 
The user can specify the names of the low and high columns in the output.
Defaults are "lo" and "hi".
}
\examples{
df1 <- dataCensored[dataCensored$station=="CB3.3C","chla"][1:30]
colnames(df1)
# Default values
df2 <- unSurv(df1)
colnames(df2)
# User values
df3 <- unSurv(df1, "LOW", "HIGH")
colnames(df3)

}
