% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unSurv.R
\name{unSurv}
\alias{unSurv}
\title{Converts Surv object into a 3-column matrix}
\usage{
unSurv(x)
}
\arguments{
\item{x}{vector (Surv object)}
}
\value{
Returns a 3-column matrix: lo, hi, type
}
\description{
Converts Surv object into a 3-column matrix
}
\details{
The type column has the following meanings:

 1  -- no censoring
 
 2  -- left censored ("less than in a survival sense" [-Inf to 10], i.e., <10)
 
 3  -- interval censored ("less than in a water quality sense", i.e., "0 - <3", "1 - 3")
 
 NA -- missing value
}
\examples{
\dontrun{
x1 <- dataCensored[dataCensored$station=="CB3.3C","chla"][1:30]
x2 <- unSurv(x1)
}

}
