% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unSurv.R
\name{unSurvDF}
\alias{unSurvDF}
\title{Converts Surv objects in a dataframe to "lo" and "hi" values}
\usage{
unSurvDF(df)
}
\arguments{
\item{df}{dataframe with Surv objects}
}
\value{
Returns dataframe with censored data converted to lo/hi format
}
\description{
Converts Surv objects in a dataframe to "lo" and "hi" values
}
\examples{
\dontrun{
df <- dataCensored[dataCensored$station=="CB3.3C", ][1:20,]
df2 <- unSurvDF(df)
}

}
