% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usgsGages}
\alias{usgsGages}
\title{USGS Gages}
\format{A data frame with 9 rows and 2 variables:
\describe{
   \item{usgsGageID}{USGS Gage ID}
   \item{siteName}{USGS Site Name}
}}
\usage{
usgsGages
}
\description{
Long Description
}
\keyword{datasets}
